+++ src/FatturaElettronicaFactory.php
@@ -12,9 +12,12 @@
 namespace Deved\FatturaElettronica;
 
 use Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaBody;
+use Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaBody\Allegato;
+use Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaBody\DatiBeniServizi\DatiRiepilogo;
 use Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaBody\DatiBeniServizi\DettaglioLinee;
 use Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaBody\DatiGenerali;
 use Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaBody\DatiPagamento;
+use Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaBody\DatiVeicoli;
 use Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaHeader;
 use Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaHeader\CedentePrestatore;
 use Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaHeader\CessionarioCommittente;
@@ -97,7 +100,7 @@ class FatturaElettronicaFactory
     }
 
     /**
-     * @param DatiTrasmissione\IdTrasmittente $idTrasmittente
+     * @param IdTrasmittente $idTrasmittente
      */
     public function setIdTrasmittente(IdTrasmittente $idTrasmittente)
     {
@@ -153,7 +156,9 @@ class FatturaElettronicaFactory
      * @param DatiPagamento $datiPagamento
      * @param DettaglioLinee $linee
      * @param bool $progessivoInvio
-     * @param FatturaElettronicaBody\DatiBeniServizi\DatiRiepilogo|null $datiRiepilogo
+     * @param DatiRiepilogo|null $datiRiepilogo
+     * @param Allegato|null $allegato
+     * @param DatiVeicoli|null $datiVeicoli
      * @return FatturaElettronica
      * @throws \Exception
      */
@@ -162,7 +167,9 @@ class FatturaElettronicaFactory
         DatiPagamento $datiPagamento = null,
         DettaglioLinee $linee,
         $progessivoInvio = false,
-        FatturaElettronicaBody\DatiBeniServizi\DatiRiepilogo $datiRiepilogo = null
+        DatiRiepilogo $datiRiepilogo = null,
+        Allegato $allegato = null,
+        DatiVeicoli $datiVeicoli = null
     ) {
         if (!$this->cessionarioCommittente) {
             throw new \Exception('Dati cessionario non presenti!');
@@ -192,7 +199,9 @@ class FatturaElettronicaFactory
         $fatturaElettronicaBody = new FatturaElettronicaBody(
             $datiGenerali,
             $datiBeniServizi,
-            $datiPagamento
+            $datiPagamento,
+            $allegato,
+            $datiVeicoli
         );
         return new FatturaElettronica($fatturaElettronicaHeader, $fatturaElettronicaBody, $this->xmlFactory);
     }