+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.10] - 2019-02-05
+### Added 
+- Aggiunta possibilità di specificare DatiSAL e DatiVeicoli, pull request #51 by manrix
+
+## [1.1.9]
+### Fixed
+- DataScadenzaPagamento facoltativa in DatiPagamento
+
+## [1.1.8]
+### Added
+- possibilità di specificare allegati (grazie manrix)
+
+## [1.1.7]
+### Fixed
+- bug fix vari
+
 ## [1.1.6]
 ### Added
 - in FatturaElettronicaFactory aggiunta possibilità di settare l'IdTrasmittente